# docs/swagger.yaml
definitions:
...
  request.ProductRequest:
    properties:
      brand:
        type: string
      category_id:
        type: integer
      description:
        type: string
      product_name:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message: {}
      statusCode:
        type: integer
    type: object
...
  utils.UpdateProduct:
    properties:
      brand:
        type: string
      category_id:
        type: integer
      description:
        type: string
      product_name:
        type: string
    type: object
...
host: localhost:9000
info:
  contact:
    email: engdaworkyismaw9@gmail.com
    name: Engdawork yismaw
  description: This is a sample server for an e-commerce platform.
  title: E-commerce API
  version: "1.0"
paths:
...
  /image/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image for a product in cloudinary
      operationId: upload-product-image
      parameters:
      - description: Product Item ID
        in: formData
        name: product_item_id
        required: true
        type: integer
      - description: product image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Upload image
      tags:
      - product image
...
  /product/{id}:
    get:
      description: Get single product by id
      operationId: get-product-by-id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get product
      tags:
      - product
  /product/create:
    post:
      consumes:
      - application/json
      description: Insert a new product
      operationId: create-product
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/request.ProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: create product
      tags:
      - product
  /product/delete/{id}:
    delete:
      description: Delete product by id
      operationId: delete-product-by-id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Delete product
      tags:
      - product
...
  /product/update/{id}:
    put:
      consumes:
      - application/json
      description: Update product by id
      operationId: update-product-by-id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/utils.UpdateProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - JWT: []
      summary: Update product
      tags:
      - product
  /products/{id}:
    get:
      description: Get a single product by id
      operationId: get-products-by-id
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get single product
      tags:
      - products
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"